🧪 E011A7 test: keep doctor memo coverage focused
